move global test functions to global file
